+++ src/Pages/Signin.tsx
@@ -1,31 +1,11 @@
-import React, { useState } from "react";
-import signinAPI from "../API/signinAPI";
+import React from "react";
+import SigninForm from "../Components/SigninForm";
 
 const Signin = () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const onSubmitForm = async (e: React.FormEvent) => {
-    e.preventDefault();
-    const [token, error] = await signinAPI(email, password);
-  };
   return (
-    <form onSubmit={onSubmitForm}>
-      <input
-        type="text"
-        value={email}
-        onChange={(e) => setEmail(e.target.value)}
-      />
-      <input
-        type="password"
-        value={password}
-        onChange={(e) => setPassword(e.target.value)}
-      />
-      <input
-        type="submit"
-        value="Sign in"
-        disabled={!email.includes("@") || password.length < 8}
-      />
-    </form>
+    <div>
+      <SigninForm submit="Sign in" type="SIGN_IN" />
+    </div>
   );
 };
